Update tokei
Changed to use /bin/sh, as it seems to work. If any one has trouble, let me known. 

Also added the -n option to use notify-send to notify of time change, the the option passed the minutes between notifications. I known, not much use at all. I just thought it was cool to be able to have an icon notification on wallpaper change :).