Added offlining files
Added the ability to download all the files in a backend. This takes quite some time, as it is 1440 files. Also, I added a 5 sec pause after each DL, as I really wouldn't want to hammer the server too much, and spread out the DL a bit :). Thus, it will take a long time, hours. So start it when you go to bed :). Tested with the backend Osaka, and it worked great. Still, report bugs, ect.