Add Seconds/nanoseconds to name
If you DL more then 1 video in a min, it will over write the last one. This changes it to the nano seconds, not likely to be the same even if DLing a lot with a script. Why this is needed is youtube, it gives it a long name that Kodi can't see, so needs  a rename, this is the best I can do :).